refactor(distance-map): introduce intermediate distance map storing source to target configuration.
